requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-ratelimit==3.0.0
pyotp==2.8.0
pillow==10.2.0
psycopg2-binary==2.9.9
django-request-id==1.0.0
django-guid==3.4.0
django-cors-headers==3.14.0
django-filter==23.1
rest-framework-roles
python-dotenv
qrcode==7.4.2
cairosvg
requests==2.31.0
django-cors-headers
pandas==2.0.2
sib-api-v3-sdk==7.6.0